seperated scipy and optimizer tests
